Upgrade to latest consul-template and migrate to hclog
Upstream consul-template now accepts a stdlib logger for its child
implementation, which allows us to move off the older fork of
consul-template we've been using. Rather than implement a shim between
logrus and the stdlib logger, this PR also updates remco to use hclog
and use its provided shim instead.

This is functionally equivalent to what we were using logrus for,
although likely not byte-for-byte compatible with the older format of
the logs.

Lastly, as part of this migration away from logrus, a couple of
exposed methods have been removed from the log module (SetLevel,
SetOutput) which were unused outside of tests (for said methods).